[FEAT] Set up the project structure
	***Disclaimer***
The whole project is expected to stay as
simple as possible. Once I reach the stage
when am I happy with how it works on my machine,
I am not going to fiddle with it just to add
functionalities that I never planned to include.

Makefile is used to spin up the binary and do
unit tests in the most obvious way. I keep the
README file up to date so that it reflects what
is going on with the program. All the stuff
is pretty much self explanatory.