[FEAT] Carriage return is better than new line
I replaced the new line at the end of each
message with a carriage return at the start
of a new one.

The functional output is pretty much the same,
but I think that this is a better idea for
clients that might want to print out the
output on new line, like Netcat, for instance.